// pub_akl-design/_ui/mylab/src/apis/model/kitModel.ts
export const KitStatus = {
  BEFORE: 'S001', // 맴버등록
  READY: 'S002', // 접수대기
  CREATED: 'S003', // 접수진행
  PROCEED: 'S004', // 접수완료
  SHIPPING: 'S005', // 키트배송
  ARRIVAL: 'S006', // 키트도착
  ANALYZING: 'S007', // 분석중
  DONE: 'S008', // 분석완료
  REJECT: 'S009', // 키트반려
  WITHDRAWAL: 'S010', // 회원탈퇴
} as const

export type KitStatus = typeof KitStatus[keyof typeof KitStatus]
export interface KitStatusRes {
  id: KitStatus
  name: string
  type: string
}